+++ crates/zng-wgt/src/lib.rs
@@ -60,9 +60,9 @@ mod __prelude {
 
     pub use zng_var::{
         ContextVar, IntoValue, IntoVar, ObservableVec, ResponderVar, ResponseVar, Var, VarCapability, VarHandle, VarHandles, VarUpdateId,
-        VarValue, WeakVar, const_var, context_var, expr_var, impl_from_and_into_var, merge_var, response_done_var, response_var, var,
-        var_default, var_from, var_state, when_var,
-    }; // TODO(breaking) add flat_expr_var!
+        VarValue, WeakVar, const_var, context_var, expr_var, flat_expr_var, impl_from_and_into_var, merge_var, response_done_var,
+        response_var, var, var_default, var_from, var_state, when_var,
+    };
 
     pub use zng_layout::{
         context::{DIRECTION_VAR, LAYOUT, LayoutDirection, LayoutMetrics},
@@ -70,8 +70,9 @@ mod __prelude {
             Align, AngleDegree, AngleGradian, AngleRadian, AngleUnits as _, ByteUnits as _, Dip, DipBox, DipPoint, DipRect, DipSideOffsets,
             DipSize, DipToPx as _, DipVector, Factor, Factor2d, FactorPercent, FactorSideOffsets, FactorUnits as _, Layout1d as _,
             Layout2d as _, LayoutAxis, Length, LengthUnits as _, Line, LineFromTuplesBuilder as _, Point, Px, PxBox, PxConstraints,
-            PxConstraints2d, PxCornerRadius, PxLine, PxPoint, PxRect, PxSideOffsets, PxSize, PxToDip as _, PxTransform, PxVector, Rect,
-            RectFromTuplesBuilder as _, ResolutionUnits as _, SideOffsets, Size, TimeUnits as _, Transform, Vector,
+            PxConstraints2d, PxCornerRadius, PxDensity, PxDensity2d, PxDensityUnits as _, PxDensityUnits, PxLine, PxPoint, PxRect,
+            PxSideOffsets, PxSize, PxToDip as _, PxTransform, PxVector, Rect, RectFromTuplesBuilder as _, SideOffsets, Size,
+            TimeUnits as _, Transform, Vector,
         },
     };